+++ CHANGES.md
@@ -3,3 +3,5 @@
 - Added a simple calculator
 ## Update 2
 - Added a temperature converter
+## Update 3
+- Added an odd or even number checker!